move driver func test
